ml-engine/requirements.txt
flask
flask-cors
ultralytics
Pillow
gunicorn
opencv-python-headless
numpy
requests